Removing the concept of `container` components and integrating redux decorating directly in component files.
