+++ .gitignore
@@ -2,7 +2,6 @@
 .DS_Store
 .env.development.local
 .env.local
-.env.local
 .env.production.local
 .env.test.local
 .pnp.js
